+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # CHANGELOG
+
 ## [Unreleased]
+
+# Pre-alpha [v0.1.0] - 2019-12-13
+### Added:
+- Added basic logging
+- Added a kivy config file
+- Added ScreenManager and the game's screens
+- Added all function signatures
+
+### Changed:
+- Update .gitignore
+- Refactored code
+- Annotated functions
+- Renamed 'enums.py' to 'constants.py'
+- Fixed _some_ minor bugs (Code still doesn't run)
+- Documented modules, classes and functions
+
 ## Pre-alpha [v0.0.0] - 2019-11-19
 ### Added:
 - Added CHANGELOG.md